Refactor to use local storage instead of state.
