+++ src/components/User/User.js
@@ -5,7 +5,8 @@ export default class User extends Component {
 
   iconChange(){
     let id = JSON.parse(localStorage.userData).id
-    return(this.props.friendReqs.outgoing.map((request) => {
+    let friendReqs = JSON.parse(localStorage.friendReqs)
+    return(friendReqs.outgoing.map((request) => {
       if(request.user_id === id && request.friend_id === this.props.user.id){
         return <i key={'friendReq' + request.id} className="fa fa-clock-o"
                   aria-hidden="true"></i>
